pathway_rag/requirements.txt
pathway[xpack-llm,xpack-llm-docs]>=0.13.0
python-dotenv>=1.0.0
openai>=1.0.0

